requirements.txt
telebot~=0.0.4
requests~=2.27.1
python-dotenv~=0.20.0
loguru~=0.6.0
peewee~=3.15.1